회원 탈퇴시 작성한 게시물 먼저 지우기

+++ src/main/java/com/example/prjjsp/service/MemberService.java
@@ -1,6 +1,7 @@
 package com.example.prjjsp.service;
 
 import com.example.prjjsp.dto.Member;
+import com.example.prjjsp.mapper.BoardMapper;
 import com.example.prjjsp.mapper.MemberMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,8 +13,8 @@ import java.util.List;
 @RequiredArgsConstructor
 @Transactional
 public class MemberService {
-
     private final MemberMapper mapper;
+    private final BoardMapper boardMapper;
 
     public void addMember(Member member) {
         mapper.insert(member);
@@ -28,7 +29,14 @@ public class MemberService {
     }
 
     public boolean remove(String id, String password) {
-        int cnt = mapper.deleteByIdAndPassword(id, password);
+        // 게시물(board) 먼저 삭제 후
+        // 회원(member) 삭제
+        int cnt = 0;
+        Member member = mapper.selectById(id);
+        if (member.getPassword().equals(password)) {
+            boardMapper.deleteByMemberId(id);
+            cnt = mapper.deleteByIdAndPassword(id, password);
+        }
         return cnt == 1;
     }
